Editable DailyWork stories
Summary: * Refactored adapters to make them all consistent. This way when all are done, maybe we can make one implementation of the AdapterViewActionListener because at this point every single one looks the same. To avoid repeating more code.

Test Plan: navigate to dailywork screen, swipe to stories, then tap on a story's state and change it.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

CC: jmsotuyo

Differential Revision: http://ph.monits.com/D2641