Fixed a nullable view on ViewHolder at MyStoriesAdapter
Summary:
There is a viewHolder called from two different methods where at one of them
it receives 3 views and at the other it receives only 2. So one of them is a
nullable view.

Test Plan: It was tested on a list where both methods were called and it crashed. So after making that view nullable it was tested again on a list where both methods were called and it worked as it should.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Subscribers: iloredom, rmorales

Differential Revision: http://ph.monits.com/D12486