Fixed bug
Summary:
If it throws an exception when the condition isn't fulfilled it causes a crash.
The problem is that on MyQueueWorkFragment you can receive a Story object on json response,
and if it happens it crashes and doesn't deserialize it.

Test Plan:
Added a task without Story to work queue from agilefant's web and then run the app.
It worked perfectly, but then I added a task with Story and then run the app
It crashed.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Subscribers: rmorales, rpereyra

Differential Revision: http://ph.monits.com/D12713