Drag expanded stories
Summary:
There was an issue where if an expanded Story had more items than the amount that could be shown (because of screen size)
we got a crash. Thats because recyclerView was trying to get a viewHolder that wasnt drawn yet.
So what I did is to request only the visible items and when you drag it draws new items automagically.
That might be an temTouchHelper.Callback thing.

What i don't like about the fix is that we re assuming that getLayoutManager() is going to return a
LinearLayoutManager.

Test Plan: Tested it's functionality by using it

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Subscribers: iloredom, rpereyra, rmorales

Differential Revision: http://ph.monits.com/D13065